style: 调整一些函数的命名和位置
